:black_square_button:Containerized the front end
This change set adds a Dockerfile which produces a production ready container
which first builds the Angular app then services it via Nginx. Secondly it
updates the docker-compose.yml file in the root to include the Angular app,
this does _not_ however use the Dockerfile as that is a compiled and ready
version of the app, what the docker-compose.yml file does is bring up a node
container and executes npm run start which starts the development server.

ng serve uses port 49153 by default to push updates thus that needs to be
exposed. Also the --poll flag is needed, but that might only need to be
used on Windows. I only tested this on Windows and it is what I needed but
the interwebs says it shouldn't be needed on other platforms.