Update expandable menu to slide out from the left
This change set updates the expandable menu, the hamberger menu when the
display is small, so that instead of sliding down from the top it pops out
from the left. It also covers the main part of the page with an overlay while
the menu is open so that it stands out more.