Add form to input katas
This change set adds a new view to the app which lets users input new Katas
into the registry. It includes updating the nav menu so one can access the
form.

It also includes adding icons to the collapsable menu for each entry.

It does not yet include authentication protection