Add a Kata list view
This change set adds a list view component which fetches all the katas from
the backend and shows them in a list. It uses a _kata card_ component which
is responsable for visualizing a Kata, a _kata service_ for communicating
with the backend for fetching the collection, and a _kata model_ for defining
what properties a Kata has.