Add Create route for posting new Katas
This change set adds a new route in the backend which allows users to POST
a new kata into the registry.

Note: I disabled the dangle underscore rule because Mongoose uses an
underscore in it's default ID property (_id)

Note: I also updated the Kata model to include a
created at / updated at field, I did this by updating the Schema
in Mongoose so that it adds those fields automatically.