:bug:Fix kata list view
The kata list view was not showing the katas because the service was expecting
the body of the GET to be an array but it is actually an object where the
array is under a property of said object.

This change set updates the service to map the array to the root level
so that the component just gets the collection vs. the response body