Add package-lock.json
This change set commits the package-lock.json file and updates the test
script to run npm install vs. npm prune/npm update.

npm update will update a package lock file but really we just want to make sure
that the tests are run with the latest versions of dependencies we are
deploything with. That is if I update a dependency I want the tests to pick
that up and update the installed dependency prior to running the tests. I
don't however want the tests to _update_ the dependencies on me.

The reason it was npm update vs. npm install is that npm install will not
update what is installed even if the package.json has been updated. If you
already have an older version installed it will keep using that version.

In Node 8 however package-lock.json was added and now when you run
npm install as long as you have a package-lock.json file it will update
what you have installed to match what is in the package.json file. I think
it also runs npm prune as it were so no need to run that command any more
either.

This will speed up the starting of the tests as this form of npm install runs
a lot faster then npm prune and npm update.