Add bench mark tests for the Kata routes
This change set adds bench mark tests for the Kata routes so we can tell
if we're making things slower.

Note: Had to add quotes around the glob in the bench NPM script as it was not
evaluating correctly. As it was the schell used by NPM (i.e. bash 3.2) was
evaluating the glob but as that version of bash does not support ** it was not
finding both *.bench.js files. By adding quotes around the glob it prevents
the shell from evaluating it and passes it as is into api-bench-runner which
uses a package to evaluate the glob correclty.