:art:Move Router components to routing module
Currently when adding a new component which can be routed to one needed to
add the route (by importing the component) to the routing module _and_ add
the component to the declarations collection in the app module.

This change set simplifies this by putting the collection of declarable
components for anything routable in the routing module thus you only have
to import them once.

With this change the app module just imports the routes and the components
from the routing module and imports/declars them. Now the app module does
not need to be changed when we add new components. The Routing module now
exports the routes _and_ the collection of declarable components. Each
time I add a new routable component I have to add the component to each
list but now I only have to touch one file.

I got this from a Youtuber (Codevolution) via their tutorial on Angular 6
routing.