:art:Refactor list view to just show Kata name
This change set removes the Kata-Card component and refactors the kata list
view to just show the kata name vs. the Kata name and description.

This way we can show more Katas above the fold on the list view and focuses
the user towards clicking on a Kata to view it.