Add route to update Katas
This change set adds a new route to allow for the updating of a Kata.
This route is attached to both the PUT and PATCH verbs, I perfer PATCH as
this route is obaying the PATCH spec but as most are using PUT/PATCH
interchangably I figured I'd hook up the PUT verb too.