:lock:Update login response to only return 401
Here I was returning 404 when the email provided during a login did not exist,
that would leak out whose emails existed in the system. I could try random
peoples emails and if I got a 401 vs. a 404 I would know they have an account
in the system.

To avoid the leak this change set updates the log in route to only return
401's regardless if the provided email exists or not