:bug:Fix Add Kata menu link so it close the menu
The slide out menu when you click on a menu item should close and remove the
overlay. This worked for all links except the Add Kata link.

That was because the add Kata link was _added_ after the ngOnInit() hook was
called. This is because it is made visiable only if the user is logged in.

The ngOnInit() hook would run jQuery which added the close menu hook to each
menu link. Since the Add Kata entry was not their yet it was missing the
close menu hook.

To fix this I switched to Event Delegation form where I attach a click event
handler to the document but tag it only for the li > a element's click
event. This way the event handler is only added once to a static element
which happens to be a parent of the anchours so their events bubble up
to the parent and get handled. The fact that some achors are added to the
parent dynamicly is no longer an issue.