Add Init Admin View
This change set updates the backend to include an /admin route which servies
up a UI that allows a user to create an Admin for the site. This view is
only accessable if no admins exist and is used to setup the first
Admin of the site. Once the admin is created navigate to the front
end and sign in normally.