Add addedBy property to Kata model
This change set updates the backend to add a new property to the Kata model
so that we can track who added a Kata. This change includes the updates
to the controller to include the new property. On inserts the addBy user
is pulled from the request. On fetch the addBy user is returned