Add a get all Users route
This change set adds a new route to the backend which retrives all the
registered users.

This route is only accessable for logged in users so if your not logged in then
you'll get a 401. Secondly the route is only accessable by site Admins so if
you are not an admin you will also get a 401.