Add Update User Route
This change set adds a new route which can be used to update a user object.

This route is only accessable by logged in users and those users must be
admin users.

This change set includes adding a new property to the User model called state
which allows users to be set to Pending or Active states. Currently there is
no logic which uses this state property but now that we have it and can
change it via this new update route we can update the app to not allow pending
users to log in until their account is _approved_.

Note: I added both PATCH and PUT verbs to this route, it is a PATCH route but
felt why not also use the PUT verb, it is not strictly a PUT but meh.