Update auth route to return full user
This change set updates the authentication route to return the entire
presented user payload. Before it was just returning the display name
and admin flag but now it will also include the state and id.

This was mostly about getting the users id back so we can do some
comparisons against Katas to see which Katas the user owns