Update auth service to store full user
This change set updates the authentication service to save the entire
presented user payload. Before it was just saving the display name
and admin flag but now it will also include the state and id.

This was mostly about saving the users id so we can do some
comparisons against Katas to see which Katas the user owns