Update Auth service to broadcast entire user object
This change set updates the Authenitcation service so that when a user logs
in instead of just broadcasting their display name it will now broadcast
the entire user payload allowing consumers to pull what information they need.

This way we can make sure that consumers can get an updated user Id when the
user changes.