Update UI to hide delete button when not my Kata
Non-admin users are not allowed to delete others Katas. This change set
updates the UI to hide the Delete/Edit buttons if the user is viewing someone
elses Kata. If however they are an admin then they can still edit/delete the
others Kata so in that case the buttons are still shown.