Add total to fetch all Kata payload
This change set updates the payload returned by the fetch all kata to include
the total number of Katas _if_ pagination is used. If no pagination queyr
parameter was set then the payload will _not_ have a total property.

Note: The unit/integration tests show that it will and have it set to
undefined but the system tests show that the payload will not have the property
that is because the act of actually JSON-ifying the object will strip any
undefined properties. The unit/integration tests forgo the actual
JSON-ification so the undefined total property remains.

The total is returned as part of the pagination API of the model however we
don't get the value with using find. To save an extra lookup for the total
and since the use of the total is for calculating the number of pages
for pagination we'll forgo looking up the total when the pagination
query parameters are missing.