Improve log messages
The logs generated by the backend were rather verbose, they were logging
both the full request and response objects. This is because we were not
using buynan's built in req/res serializers so the entire objects were
being serialized and put into the log message. The req and res objects
are added by the express-buynan-logger.

This change set updates the buynan logger configuration to enable the use
of the req/res serializers so that buynan will strip those objects down
to a few properties instead of the entire objects. This will reduce the
size of each log entry and make the logs much easier to read