:shirt: disable func-names in user model
The user model file has a number of linter errors regarding unamed functions.

The User model is a Mongoos model and in that file we define custom
methods and functions which need to be actual functions as they will be
using the calling object via the 'this' keyword. Using an arrow function here
would not work so these linter errors can be ignored.

This change set just disables the rule for this file so they no longer
show up in the linter report